refactor: start rewiring things
