+++ packages/ui/src/index.ts
@@ -2,4 +2,4 @@ export { Button } from './button';
 export type { ButtonProps } from './button';
 
 export { Modal } from './components/Modal';
-export { Portal } from './components/Portal';
+export { Portal } from './components/PortalOld';
